# Conflicts:
#	examples/opensora_hpcai/opensora/models/layers/blocks.py
#	examples/opensora_hpcai/opensora/models/stdit/stdit2.py

+++ examples/diffusers/text_to_image/train_text_to_image_lora_sdxl.py
@@ -32,8 +32,8 @@ from tqdm.auto import tqdm
 from transformers import AutoTokenizer, PretrainedConfig
 
 import mindspore as ms
-from mindspore import context, nn, ops
-from mindspore.amp import DynamicLossScaler, LossScaler, StaticLossScaler, all_finite
+from mindspore import nn, ops
+from mindspore.amp import StaticLossScaler
 from mindspore.dataset import GeneratorDataset, transforms, vision
 
 from mindone.diffusers import AutoencoderKL, DDPMScheduler, StableDiffusionXLPipeline, UNet2DConditionModel
@@ -43,6 +43,8 @@ from mindone.diffusers._peft.utils import get_peft_model_state_dict
 from mindone.diffusers.loaders import LoraLoaderMixin
 from mindone.diffusers.optimization import get_scheduler
 from mindone.diffusers.training_utils import (
+    AttrJitWrapper,
+    TrainStep,
     _set_state_dict_into_text_encoder,
     cast_training_params,
     compute_snr,
@@ -401,7 +403,6 @@ def parse_args(input_args=None):
     def error_template(feature, flag):
         return f"{feature} is not yet supported, please do not set --{flag}"
 
-    assert args.gradient_accumulation_steps == 1, error_template("Gradient Accumulation", "gradient_accumulation_steps")
     assert args.allow_tf32 is False, error_template("TF32 Data Type", "allow_tf32")
     assert args.use_8bit_adam is False, error_template("AdamW8bit", "use_8bit_adam")
     if args.push_to_hub is True:
@@ -830,8 +831,8 @@ def main():
     lr_scheduler = get_scheduler(
         args.lr_scheduler,
         args.learning_rate,
-        num_warmup_steps=args.lr_warmup_steps * args.gradient_accumulation_steps,
-        num_training_steps=args.max_train_steps * args.gradient_accumulation_steps,
+        num_warmup_steps=args.lr_warmup_steps,
+        num_training_steps=args.max_train_steps,
     )
     # Optimizer creation
     # we have to add prefix to param.name, otherwise the optimizer gives a fucking param.name duplication error!
@@ -885,16 +886,16 @@ def main():
         else:
             logger.warning(f"Tracker {tracker_name} is not implemented, omitting...")
 
-    train_step = TrainStep(
+    train_step = TrainStepForSDXLLoRA(
         vae=vae,
         text_encoder_one=text_encoder_one,
         text_encoder_two=text_encoder_two,
         unet=unet,
         optimizer=optimizer,
-        scaler=StaticLossScaler(65536),
         noise_scheduler=noise_scheduler,
-        args=args,
         weight_dtype=weight_dtype,
+        length_of_dataloader=len(train_dataloader),
+        args=args,
     ).set_train()
 
     if args.enable_mindspore_data_sink:
@@ -977,51 +978,55 @@ def main():
         if args.train_text_encoder:
             text_encoder_one.set_train(True)
             text_encoder_two.set_train(True)
+        train_loss = 0.0
         for step, batch in (
             ((_, None) for _ in range(len(train_dataloader)))  # dummy iterator
             if args.enable_mindspore_data_sink
             else enumerate(train_dataloader_iter)
         ):
             if args.enable_mindspore_data_sink:
-                loss = sink_process()
+                loss, model_pred = sink_process()
             else:
-                loss = train_step(*batch)
+                loss, model_pred = train_step(*batch)
+            train_loss += loss.numpy().item()
 
-            progress_bar.update(1)
-            global_step += 1
-            for tracker_name, tracker in trackers.items():
-                if tracker_name == "tensorboard":
-                    tracker.add_scalar("train/loss", loss.numpy().item(), global_step)
+            if train_step.sync_gradients:
+                progress_bar.update(1)
+                global_step += 1
+                for tracker_name, tracker in trackers.items():
+                    if tracker_name == "tensorboard":
+                        tracker.add_scalar("train/loss", train_loss, global_step)
+                train_loss = 0.0
 
-            if is_master(args):
-                if global_step % args.checkpointing_steps == 0:
-                    # _before_ saving state, check if this save would set us over the `checkpoints_total_limit`
-                    if args.checkpoints_total_limit is not None:
-                        checkpoints = os.listdir(args.output_dir)
-                        checkpoints = [d for d in checkpoints if d.startswith("checkpoint")]
-                        checkpoints = sorted(checkpoints, key=lambda x: int(x.split("-")[1]))
+                if is_master(args):
+                    if global_step % args.checkpointing_steps == 0:
+                        # _before_ saving state, check if this save would set us over the `checkpoints_total_limit`
+                        if args.checkpoints_total_limit is not None:
+                            checkpoints = os.listdir(args.output_dir)
+                            checkpoints = [d for d in checkpoints if d.startswith("checkpoint")]
+                            checkpoints = sorted(checkpoints, key=lambda x: int(x.split("-")[1]))
 
-                        # before we save the new checkpoint, we need to have at _most_ `checkpoints_total_limit - 1` checkpoints
-                        if len(checkpoints) >= args.checkpoints_total_limit:
-                            num_to_remove = len(checkpoints) - args.checkpoints_total_limit + 1
-                            removing_checkpoints = checkpoints[0:num_to_remove]
+                            # before we save the new checkpoint, we need to have at _most_ `checkpoints_total_limit - 1` checkpoints
+                            if len(checkpoints) >= args.checkpoints_total_limit:
+                                num_to_remove = len(checkpoints) - args.checkpoints_total_limit + 1
+                                removing_checkpoints = checkpoints[0:num_to_remove]
 
-                            logger.info(
-                                f"{len(checkpoints)} checkpoints already exist, removing {len(removing_checkpoints)} checkpoints"
-                            )
-                            logger.info(f"removing checkpoints: {', '.join(removing_checkpoints)}")
+                                logger.info(
+                                    f"{len(checkpoints)} checkpoints already exist, removing {len(removing_checkpoints)} checkpoints"
+                                )
+                                logger.info(f"removing checkpoints: {', '.join(removing_checkpoints)}")
 
-                            for removing_checkpoint in removing_checkpoints:
-                                removing_checkpoint = os.path.join(args.output_dir, removing_checkpoint)
-                                shutil.rmtree(removing_checkpoint)
+                                for removing_checkpoint in removing_checkpoints:
+                                    removing_checkpoint = os.path.join(args.output_dir, removing_checkpoint)
+                                    shutil.rmtree(removing_checkpoint)
 
-                    save_path = os.path.join(args.output_dir, f"checkpoint-{global_step}")
-                    # TODO: save optimizer & grad scaler etc. like accelerator.save_state
-                    os.makedirs(save_path, exist_ok=True)
-                    save_model_hook(models, save_path)
-                    output_model_file = os.path.join(save_path, "pytorch_model.ckpt")
-                    ms.save_checkpoint(unet, output_model_file)
-                    logger.info(f"Saved state to {save_path}")
+                        save_path = os.path.join(args.output_dir, f"checkpoint-{global_step}")
+                        # TODO: save optimizer & grad scaler etc. like accelerator.save_state
+                        os.makedirs(save_path, exist_ok=True)
+                        save_model_hook(models, save_path)
+                        output_model_file = os.path.join(save_path, "pytorch_model.ckpt")
+                        ms.save_checkpoint(unet, output_model_file)
+                        logger.info(f"Saved state to {save_path}")
 
             logs = {"step_loss": loss.numpy().item(), "lr": optimizer.get_lr().numpy().item()}
             progress_bar.set_postfix(**logs)
@@ -1066,7 +1071,7 @@ def main():
             tracker.close()
 
 
-class TrainStep(nn.Cell):
+class TrainStepForSDXLLoRA(TrainStep):
     def __init__(
         self,
         vae: nn.Cell,
@@ -1074,41 +1079,20 @@ class TrainStep(nn.Cell):
         text_encoder_two: nn.Cell,
         unet: nn.Cell,
         optimizer: nn.Optimizer,
-        scaler: LossScaler,
         noise_scheduler,
-        args,
         weight_dtype,
+        length_of_dataloader,
+        args,
     ):
-        super().__init__()
-        self.unet = unet.set_grad()
-        self.optimizer = optimizer
-        self.weights = optimizer.parameters
-        self.scaler = scaler
-        if isinstance(self.scaler, StaticLossScaler):
-            self.drop_overflow = False
-        elif isinstance(self.scaler, DynamicLossScaler):
-            self.drop_overflow = True
-        else:
-            raise NotImplementedError(f"Unsupported scaler: {type(self.scaler)}")
-        self.parallel_mode = context.get_auto_parallel_context("parallel_mode")
-        if self.parallel_mode == context.ParallelMode.STAND_ALONE:
-            self.grad_reducer = nn.Identity()
-        elif self.parallel_mode in (context.ParallelMode.DATA_PARALLEL, context.ParallelMode.HYBRID_PARALLEL):
-            self.grad_reducer = nn.DistributedGradReducer(self.weights)
-        else:
-            raise NotImplementedError(f"When creating reducer, Got Unsupported parallel mode: {self.parallel_mode}")
-        if isinstance(unet, nn.Cell) and unet.jit_config_dict:
-            self._jit_config_dict = unet.jit_config_dict
-        self.clip_grad = args.max_grad_norm is not None
-        self.clip_value = args.max_grad_norm
-
-        @ms.jit_class
-        class ArgsJitWrapper:
-            def __init__(self, **kwargs):
-                for name in kwargs:
-                    setattr(self, name, kwargs[name])
-
-        self.args = ArgsJitWrapper(**vars(args))
+        super().__init__(
+            unet,
+            optimizer,
+            StaticLossScaler(65536),
+            args.max_grad_norm,
+            args.gradient_accumulation_steps,
+            gradient_accumulation_kwargs=dict(length_of_dataloader=length_of_dataloader),
+        )
+        self.unet = self.model
         self.vae = vae
         self.vae_scaling_factor = self.vae.config.scaling_factor
         self.text_encoder_one = text_encoder_one
@@ -1117,8 +1101,7 @@ class TrainStep(nn.Cell):
         self.noise_scheduler_num_train_timesteps = noise_scheduler.config.num_train_timesteps
         self.noise_scheduler_prediction_type = noise_scheduler.config.prediction_type
         self.weight_dtype = weight_dtype
-
-        self.forward_and_backward = ops.value_and_grad(self.forward, None, weights=self.weights, has_aux=True)
+        self.args = AttrJitWrapper(**vars(args))
 
     def forward(self, pixel_values, input_ids_one, input_ids_two, add_time_ids):
         # Convert images to latent space
@@ -1182,31 +1165,8 @@ class TrainStep(nn.Cell):
             loss = loss.mean(axis=list(range(1, len(loss.shape)))) * mse_loss_weights
             loss = loss.mean()
 
-        loss = self.scaler.scale(loss)
+        loss = self.scale_loss(loss)
         return loss, model_pred
-
-    def update(self, loss, grads):
-        if self.clip_grad:
-            loss = ops.depend(loss, self.optimizer(ops.clip_by_global_norm(grads, clip_norm=self.clip_value)))
-        else:
-            loss = ops.depend(loss, self.optimizer(grads))
-        return loss
-
-    def construct(self, *inputs):
-        (loss, model_pred), grads = self.forward_and_backward(*inputs)
-        grads = self.grad_reducer(grads)
-        loss = self.scaler.unscale(loss)
-        grads = self.scaler.unscale(grads)
-
-        if self.drop_overflow:
-            status = all_finite(grads)
-            if status:
-                loss = self.update(loss, grads)
-            loss = ops.depend(loss, self.scaler.adjust(status))
-        else:
-            loss = self.update(loss, grads)
-
-        return loss
 
 
 def validate(pipeline, args, trackers, logging_dir, epoch):
